CMSC335 Project Added; Repo and Code Refactoring
Content Change: Added the CMSC335 project to the website, along with the video.
we made at McKeldin.

Made a commitment to a static personal website, so removed the Dockerfile and
docker-compose.yml. If I want to add dynamic functionality in the future, in
would have to be through calling another server with client-side javascript.

Additionally, I split out the javascript from the website so that it can be used
in multple files.

Finally, I made new media and docs folder, separate and organize the website
code, from the fonts, images, videos, papers and other documents.