Updated Maria Video and Added Papers
Updated the Maria video to be a direct file. This helps, because we were
previously using an embedded YouTube iframe, which slowed down the website, and
made a lot of spurious requests to Google services. This video directly on the
website should save 0.25 seconds on load time.

Also, I am no longer using the Salsa font in my page. This also speeds up the
page a good amount.

Finally, I added my philosophy papers from 2021 to the website. I don't know how
wise it is seeing as looking back on them they aren't actually that
groundbreaking (I mean hey, I was 18 at the time), but none-the-less, they are
now on the site.