README.md and Description Tag
Added a README.md to website repo, and added description tags to both index.html
and qr.html. These don't do anything for the website on the outside, but they do
help with SEO. Right now the descriptions when you search up the site kinda
suck, so this should help with that.