Added the Report on the Netherlands
Added the Report on the Netherlands from the School Work repo. Note it is not
the same .html file as there. It is in the format of the website, and it has
been polished in some ways. The article still requires some polish, but is in a
mostly working state for the purposes of the trip. In the future the writing
could be improved, and links to read more about each section could be added to
the respective sections.