Instagram Style Guide
Copied over the Instagram Style Guide from the School Work repository. It's not
yet public (I haven't put a link in the index.html), because I want to improve
it more before posting publically. I want to aim for a public post around the
time of the Trumpet House post, which will be June. 29, by the look of it